requirements.txt
pylint == 2.12.2
pytest ==6.2.5
pytest-html ==3.1.1
pymysql==1.0.2
flask ==2.0.2
flask-sqlalchemy ==2.5.1
flask-marshmallow == 0.14.0
marshmallow-sqlalchemy == 0.27.0
Werkzeug~=2.0.2
six~=1.16.0
pytz~=2021.3
setuptools==60.5.0
aniso8601~=9.0.1
flask-wtf==1.0.0
flask-restful==0.3.9
mysql-connector-python==8.0.27
mysql-connector-python-rf==2.2.2
requests==2.27.1
gunicorn==20.1.0